requirements.txt
dash==2.16.1
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
smtplib-ssl==1.0.0
dash-bootstrap-components==1.5.0
schedule==1.2.0
pyyaml==6.0.1
cryptography==41.0.8
